requirements.txt
openai>=1.0.0
pillow>=10.0.0
moviepy>=1.0.3
gtts>=2.4.0
pdf2image>=1.16.3
pytesseract>=0.3.10
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
matplotlib>=3.7.0
sympy>=1.12
